+++ spec/lib/slack/block_kit/element/timepicker_spec.rb
@@ -3,21 +3,10 @@
 require 'spec_helper'
 
 RSpec.describe Slack::BlockKit::Element::Timepicker do
-  let(:instance) { described_class.new(action_id: action_id) }
-  let(:type) { 'timepicker' }
-  let(:placeholder) { 'some text' }
-  let(:initial_time) { '11:23' }
-  let(:action_id) { '1123' }
-  let(:placeholder_json) do
-    Slack::BlockKit::Composition::PlainText.new(
-      text: placeholder
-    ).as_json
-  end
-
   describe '.initialize' do
     it 'yields self' do
       yielded = nil
-      new_instance = described_class.new(action_id: action_id) do |timepicker|
+      new_instance = described_class.new(action_id: '1123') do |timepicker|
         yielded = timepicker
       end
 
@@ -25,24 +14,14 @@ RSpec.describe Slack::BlockKit::Element::Timepicker do
     end
   end
 
-  describe '#placeholder' do
-    it 'returns self' do
-      expect(instance.placeholder(text: placeholder)).to be(instance)
-    end
-  end
-
-  describe '#initial_time' do
-    it 'returns self' do
-      expect(instance.initial_time(initial_time)).to be(instance)
-    end
-  end
-
   describe '#as_json' do
     context 'when minimal use case' do
       it 'encodes type and action_id' do
+        instance = described_class.new(action_id: '1123')
+
         expect(instance.as_json).to eq(
-          type: type,
-          action_id: action_id
+          type: 'timepicker',
+          action_id: '1123'
         )
       end
     end
@@ -50,14 +29,17 @@ RSpec.describe Slack::BlockKit::Element::Timepicker do
     context 'with a placeholder' do
       let(:expected_json) do
         {
-          type: type,
-          action_id: action_id,
-          placeholder: placeholder_json
+          type: 'timepicker',
+          action_id: '1123',
+          placeholder: Slack::BlockKit::Composition::PlainText.new(text: 'pick a time').as_json
         }
       end
 
       it 'encodes the placeholder object' do
-        instance.placeholder(text: placeholder)
+        instance = described_class.new(
+          action_id: '1123',
+          placeholder: 'pick a time'
+        )
 
         expect(instance.as_json).to eq(expected_json)
       end
@@ -66,14 +48,17 @@ RSpec.describe Slack::BlockKit::Element::Timepicker do
     context 'with initial_time' do
       let(:expected_json) do
         {
-          type: type,
-          action_id: action_id,
-          initial_time: initial_time
+          type: 'timepicker',
+          action_id: '1123',
+          initial_time: '11:23'
         }
       end
 
       it 'encodes the initial_time' do
-        instance.initial_time(initial_time)
+        instance = described_class.new(
+          action_id: '1123',
+          initial: '11:23'
+        )
 
         expect(instance.as_json).to eq(expected_json)
       end
@@ -82,16 +67,19 @@ RSpec.describe Slack::BlockKit::Element::Timepicker do
     context 'with placeholder and initial_time' do
       let(:expected_json) do
         {
-          type: type,
-          action_id: action_id,
-          initial_time: initial_time,
-          placeholder: placeholder_json
+          type: 'timepicker',
+          action_id: '1123',
+          initial_time: '11:23',
+          placeholder: Slack::BlockKit::Composition::PlainText.new(text: 'pick a time').as_json
         }
       end
 
       it 'encodes the placeholder & initial_time' do
-        instance.placeholder(text: placeholder)
-        instance.initial_time(initial_time)
+        instance = described_class.new(
+          action_id: '1123',
+          initial: '11:23',
+          placeholder: 'pick a time'
+        )
 
         expect(instance.as_json).to eq(expected_json)
       end